Refactor adapter with base adapter
