+++ lib/kindle_manager/adapters/books_adapter.rb
@@ -1,21 +1,5 @@
 module KindleManager
-  class BooksAdapter
-    include AmazonAuth::CommonExtension
-
-    attr_accessor :store, :session
-
-    def initialize(options)
-      @options = options
-      @session = options.fetch(:session, nil)
-      extend(AmazonAuth::SessionExtension)
-
-      @limit = options.fetch(:limit, nil)
-      @max_scroll_attempts = options.fetch(:max_scroll_attempts, 20)
-
-      @store = KindleManager::FileStore.new(options.merge(session: @session))
-      log "Directory for downloaded pages is #{store.base_dir}"
-    end
-
+  class BooksAdapter < BaseAdapter
     def fetch
       go_to_kindle_management_page
       begin
